requirements.txt
fastapi[all]
sqlmodel
psycopg2
python-multipart
starlette
pydantic
uvicorn
PyJWT~=2.8.0
cryptography
requests
SQLAlchemy
numpy